Renderer Architecture Change
