+++ Engine/src/Core/Window.cpp
@@ -9,7 +9,6 @@
 namespace aero3d {
 
 SDL_Window* Window::s_Window = nullptr;
-Scope<GraphicsContext> Window::s_Context = nullptr;
 
 bool Window::Init(const char* title, int width, int height, const char* api)
 {
@@ -35,9 +34,6 @@ bool Window::Init(const char* title, int width, int height, const char* api)
         return false;
     }
 
-    s_Context = GraphicsContext::Create(api);
-    A3D_CHECK_INIT(s_Context->Init(s_Window, width, height));
-
     return true;
 }
 
@@ -48,8 +44,6 @@ void Window::Shutdown()
     if (s_Window) {
         SDL_DestroyWindow(s_Window);
     }
-
-    A3D_SHUTDOWN(s_Context);
 
     SDL_QuitSubSystem(SDL_INIT_VIDEO);
 }
@@ -86,11 +80,6 @@ void Window::PollEvents(bool& running, bool& minimized)
         }
         }
     }
-}
-
-void Window::SwapBuffers()
-{
-    s_Context->SwapBuffers();
 }
 
 SDL_Window* Window::GetSDLWindow()
