+++ Engine/src/Core/Window.h
@@ -6,7 +6,6 @@
 #include "SDL3/SDL.h"
 
 #include "Utils/Common.h"
-#include "Graphics/GraphicsContext.h"
 
 namespace aero3d {
 
@@ -17,13 +16,11 @@ public:
     static void Shutdown();
 
     static void PollEvents(bool& running, bool& minimized);
-    static void SwapBuffers();
 
     static SDL_Window* GetSDLWindow();
 
 private:
     static SDL_Window* s_Window;
-    static Scope<GraphicsContext> s_Context;
 
 };
