--- Engine/src/Platform/Vulkan/VulkanContext.cpp
@@ -1,30 +1,0 @@
-#include "Platform/Vulkan/VulkanContext.h"
-
-#include "Utils/Log.h"
-#include "Utils/Common.h"
-#include "Platform/Vulkan/Internal/VulkanCore.h"
-
-namespace aero3d {
-
-bool VulkanContext::Init(SDL_Window* window, int width, int height)
-{
-    LogMsg("Graphics Context Initialize.");
-
-    A3D_CHECK_INIT(g_VulkanCore->Init(window, width, height));
-
-    return true;
-}
-
-void VulkanContext::Shutdown()
-{
-    LogMsg("Graphics Context Shutdown.");
-
-    A3D_SHUTDOWN(g_VulkanCore);
-}
-
-void VulkanContext::SwapBuffers()
-{
-    g_VulkanCore->SwapBuffers();
-}
-
-} // namespace aero3d
