+++ Engine/src/Platform/Vulkan/VulkanRenderer.cpp
@@ -1,0 +1,25 @@
+#include "Platform/Vulkan/VulkanRenderer.h"
+
+namespace aero3d {
+
+bool VulkanRenderer::Init()
+{
+    return true;
+}
+
+void VulkanRenderer::Shutdown()
+{
+
+}
+
+void VulkanRenderer::SubmitCommand(RenderCommand command)
+{
+    
+}
+
+void VulkanRenderer::ResizeBuffers()
+{
+
+}
+
+} // namespace aero3d
